Move file_md5sum to utils
